Relax uv_build requirements
0.8.13 [1] seems to be the last version were anything related to
uv_build was changed.

0.9.0 [2] did not introduce any breaking changes. Because the uv_build seems
to be tied to the uv version for no apparent reason and I do not
believe it to be sensible to have to update the upper limit whenever a
breaking uv and not a breaking uv_build release is made, I will be
using an open-ended version requirement and fix problems with new
versions as they arise (which I would have to do anyway eventually).

Closes #31

[1]: https://github.com/astral-sh/uv/releases/tag/0.8.13
[2]: https://github.com/astral-sh/uv/releases/tag/0.9.0